requirements.txt
appnope==0.1.0
decorator==4.0.10
Django==1.10.4
django-filter==1.0.1
djangorestframework==3.5.3
ipython==5.1.0
ipython-genutils==0.1.0
Markdown==2.6.7
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
traitlets==4.3.1
wcwidth==0.1.7
